172222087 - Created a module for converting different units of measure
Each category (weight, distance etc.) has a very small default storage unit.

Every time a conversion is made it uses the smallest default unit to convert to and from other units. e.g. to convert stones to kilograms it converts stones to grams, and then converts the grams to kilograms.

This means we don't have to code in all of the conversion formula.

It also allows for us to store 1 consistent unit and do the conversion on the fly.