172242540 - Created a module for categorizing units of measure
Units can only be converted to specific other units.
For example you can convert milliseconds to seconds but not  to yards.

This module categorizes all units into groups that can be converted to one another. It also provides a default 'stored' unit. This is the unit that the number stored in the database will represent.